fix(plugins): drop duplicate window-service re-exports and type the native window results (#926)
